@@ -28,32 +28,21 @@ include = [
 [tool.poetry.dependencies]
 python = "^3.8"
 kaleido = "0.2.1"
-pandas = "^1.0.0"
+pandas = "^2.0.0"
 plotly = "^5.3.1"
 
 [tool.poetry.group.dev.dependencies]
-autoflake = "^1.5.3"
-black = "^22.3.0"
+black = "^23.0.0"
 esbonio = "^0.13.0"
-flake8 = "^5.0.0"
-flake8-annotations = "^2.9.1"
-flake8-bandit = "^4.1.1"
-flake8-black = "^0.3.3"
-flake8-bugbear = "^22.8.23"
-flake8-builtins = "^1.5.3"
-flake8-comprehensions = "^3.10.0"
-flake8-docstrings = "^1.6.0"
-flake8-eradicate = "^1.3.0"
-isort = "^5.10.1"
-mypy = "^0.960"
-nox = "^2022.8.7"
-pep8-naming = "^0.13.2"
+mypy = "^1.0.0"
 pre-commit = "^2.19.0"
 python-kacl = "^0.2.30"
+ruff = "^0.0.259"
 
 [tool.poetry.group.test.dependencies]
+nox = "^2022.8.7"
 pytest = "^7.1.2"
-pytest-cov = "^3.0.0"
+pytest-cov = "^4.0.0"
 coverage = {extras = ["toml"], version = "^6.4.4"}
 ipykernel = "^6.6.0"
 nbformat = "^5.1.3"
@@ -79,13 +68,74 @@ exclude_lines = [
     'pragma: no cover'
 ]
 
-[tool.black]
-target-version = ["py39", "py310"]
-include = '\.pyi?$'
+[tool.ruff]
+select = [
+    "F",        # Pyflakes
+    "E", "W",   # Pycodestyle
+    "I",        # isort
+    "N",        # pep8-naming
+    "D",        # docstring
+    "UP",       # pyupgrade
+    "ANN",      # flake8-annotations
+    "B",        # flake8-bugbear
+    "A",        # flake8-builtins
+    "C40",      # flake8-comprehensions
+    "ARG",      # flake8-unused-arguments
+    "PTH",      # flake8-use-pathlib
+    "ERA",      # flake8-eradicate
+    "SIM",      # flake8-simplify
+]
 
-[tool.isort]
-profile = "black"
-src_paths = ["src", "tests", "./noxfile.py"]
+# Allow autofix for all enabled rules (when '--fix') is provided.
+fixable = [
+    "F",        # Pyflakes
+    "E", "W",   # Pycodestyle
+    "I",        # isort
+    "N",        # pep8-naming
+    "D",        # docstring
+    "UP",       # pyupgrade
+    "ANN",      # flake8-annotations
+    "B",        # flake8-bugbear
+    "A",        # flake8-builtins
+    "C40",      # flake8-comprehensions
+    "ARG",      # flake8-unused-arguments
+    "PTH",      # flake8-use-pathlib
+    "ERA",      # flake8-eradicate
+    "SIM",      # flake8-simplify
+]
+
+unfixable = []
+
+# Exclude a variety of commonly ignored directiories
+exclude = [
+    ".eggs",
+    ".git",
+    ".mypy_cache",
+    ".nox",
+    "pants.d",
+    ".ruff_cache",
+    ".venv",
+    "__pypackages__",
+    "_build",
+    "dist",
+    "site",
+]
+
+# Same as Black
+line-length = 88
+
+target-version = "py311"
+
+[tool.ruff.mccabe]
+max-complexity = 10
+
+[tool.ruff.pydocstyle]
+convention = "numpy"  # Accepts: "google", "numpy", or "pep257"
+
+[tool.black]
+target-version = ["py311"]
+include ='\.pyi?$'
+force-exclude = ".venv, .nox, docs"
 
 [tool.pytest.ini_options]
 minversion = "7.0"
@@ -95,7 +145,9 @@ addopts = """\
     --cov-report term-missing \
     --no-cov-on-fail \
     --strict-markers
+    --doctest-modules
 """
+doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL NUMBER"
 
 [tool.mypy]
 ignore_missing_imports = true
